Move HttpBlockDisk to formats
